Reorganized packages and privacy metrics
