+++ capstone/src/main/java/com/google/univiz/scorecard/CollegeIdReaderProviderImpl.java
@@ -5,7 +5,7 @@ import java.io.InputStream;
 import java.net.MalformedURLException;
 import java.net.URL;
 
-public class CollegeIdReaderProviderImpl implements CollegeIdReaderProvider {
+class CollegeIdReaderProviderImpl implements CollegeIdReaderProvider {
   @Override
   public InputStream getStreamFromUrl(String urlString) throws IOException {
     try {
