+++ capstone/src/main/java/com/google/univiz/scorecard/SuggestionDataApi.java
@@ -1,4 +1,4 @@
-package com.google.univiz.api;
+package com.google.univiz.scorecard;
 
 import com.google.univiz.api.representation.SuggestionResponse;
 import java.io.IOException;
