+++ capstone/src/main/java/com/google/univiz/scorecard/URLProvider.java
@@ -4,7 +4,7 @@ import com.google.univiz.api.representation.CollegeId;
 import java.util.List;
 
 /** Construct a URL string based on the user's preferred colleges. */
-public interface URLProvider {
+interface URLProvider {
   /**
    * Return URL in String form to a remote resource which is either aiming to get college name
    * suggestions or data resources
