+++ capstone/src/test/resources/com/google/univiz/scorecard/suggestion_multiple.json
