+++ capstone/src/test/resources/com/google/univiz/scorecard/suggestion_single.json
